requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.8.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.0.6
django-debug-toolbar==4.4.2
django-plotly-dash==2.3.1
django-redis==5.4.0
dpd_components==0.1.0
Flask==3.0.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.0.0
packaging==24.1
pandas==2.2.2
plotly==5.22.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.6
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tenacity==8.4.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
zope.interface==6.4.post2
